SrtuctredFilename: Capitalize titles properly
