+++ src/Snowflake.Tooling.Shiragame/RegionParser.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Snowflake.Romfile;
-
+using Snowflake.Romfile.Extensions;
 
 namespace Shiragame.Builder
 {
